+++ AddOns/EditModeExpanded/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Edit Mode Expanded
 
-## [10.2-012](https://github.com/teelolws/EditModeExpanded/tree/10.2-012) (2024-01-19)
-[Full Changelog](https://github.com/teelolws/EditModeExpanded/compare/10.2-011...10.2-012) [Previous Releases](https://github.com/teelolws/EditModeExpanded/releases)
+## [10.2-014](https://github.com/teelolws/EditModeExpanded/tree/10.2-014) (2024-01-23)
+[Full Changelog](https://github.com/teelolws/EditModeExpanded/compare/10.2-013...10.2-014) [Previous Releases](https://github.com/teelolws/EditModeExpanded/releases)
 
-- Fix lua error  
+- Add option to hide the players level  
+- Add option to hide the macro name and keybind name from action bar buttons  
+- Update EditModeExpanded-1.0.lua